completely rewrite how plugin and bot are communicating
this defines a bot api and a plugin api (hooks) over which
the bot and the plugin can communicate. This reduces the
complexions of the plugin massively and makes it much easier
to test and mock both sides. For now only the Slack Bot (now denoted
as platform) is ported to this new structure.

When all bots are ported to be platforms the "plugins" directory can be
deleted as all plugins will then be contained in "plugin".